God says: Let there be an update
Small QoL improvement. Changed the way admin is determined to make it so you're admin in local hosted preview (now you won't have to add yourself to admin technically)
Also changed it so that everyone can use all chat command (technically) and you just get a message that you're not certified if you're not added to the admin list.